+++ src/app/Playground/VetleSamples/TimePlanningViewModel.cs
@@ -24,6 +24,7 @@ public class TimePlanningViewModel : ViewModel
     }
 
     public List<object> Tests { get; set; } = ["Test", "Test2"];
+    public string Test { get; } = "Test";
 
     public void Update()
     {