+++ src/app/Playground/VetleSamples/ViewCellTest.xaml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using DIPS.Mobile.UI.Components.ChipGroup;
 using DIPS.Mobile.UI.Components.Pickers.ItemPicker;
 
 namespace Playground.VetleSamples;
@@ -45,9 +46,15 @@ public partial class ViewCellTest : ContentView
         if(sender is not ItemPicker itemPicker)
             return;
 
-        if (itemPicker.SelectedItem == "Test2")
-        {
-            timePlanningViewModel.Update();
-        }
+        
+    }
+
+    private async void DateAndRoundPicker_OnOnSelectedItemsChanged(object sender, ChipGroupEventArgs e)
+    {
+        if(BindingContext is not TimePlanningViewModel timePlanningViewModel)
+            return;
+        
+        await Task.Delay(1000);
+        timePlanningViewModel.Update();
     }
 }